UserService modified
Modified to use dependency Injection in UserService
